organize reusable components used by LegalView
