+++ src/components/layout/FullTextTemplate.vue
@@ -29,7 +29,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 .full-text-template__content {
   text-align: left;
 }
